Refactor: Complete architectural overhaul for modularity and maintainability
This commit finalizes a comprehensive, two-stage refactoring of the entire project, transforming it from a collection of scripts into a robust, scalable, and maintainable Python application. The new architecture is based on a clear separation of concerns.

Key Architectural Changes:
- **`src` Layout:** All source code is now centralized in the `src/` directory.
- **`core` Package:** A new `src/core` package has been created to house all shared logic:
  - `core/config.py`: A centralized module for loading and validating all environment variables.
  - `core/db/`: A dedicated package for all database interactions (`data_access.py`, `utils.py`).
  - `core/tasks.py`: All Celery background tasks are now part of the core logic.
  - `core/utils.py`: Contains shared utility functions like `rate_limit`.
- **Modular Bot Packages:** Both `library_bot` and `admin_bot` have been decomposed into their own distinct packages (`src/library_bot/`, `src/admin_bot/`) with a consistent internal structure:
  - `main.py` as the entry point.
  - `handlers/` for separating conversation logic.
  - `keyboards.py` for UI components.
  - `states.py` using Enums for robust state management.
- **Configuration & Imports:** All modules have been updated to use the new `core.config` module and absolute imports, eliminating scattered `os.getenv` calls and ensuring consistency.
- **Test Suite:** The entire test suite has been updated to align with the new architecture.

UX and Bug Fixes:
- Corrected the placement of the "Forgot Password?" button to the password entry step.
- Improved the book browsing experience by adding direct selection buttons in the genre view.
