+++ compose.yaml
@@ -26,7 +26,7 @@ services:
   library_bot:
     build: .
     restart: always
-    command: python src/librarybot.py
+    command: python src/library_bot/main.py
     volumes:
       - .:/app
     environment:
@@ -43,7 +43,7 @@ services:
   admin_bot:
     build: .
     restart: always
-    command: python src/admin_bot.py
+    command: python src/admin_bot/main.py
     volumes:
       - .:/app
     environment:
@@ -90,7 +90,7 @@ services:
   celery_worker:
     build: .
     restart: always
-    command: celery -A src.tasks.celery_app worker --loglevel=info
+    command: celery -A src.core.tasks.celery_app worker --loglevel=info
     volumes:
       - .:/app
     environment:
@@ -105,7 +105,7 @@ services:
   celery_beat:
     build: .
     restart: always
-    command: celery -A src.tasks.celery_app beat --loglevel=info
+    command: celery -A src.core.tasks.celery_app beat --loglevel=info
     volumes:
       - .:/app
     environment:
