+++ src/library_bot/handlers/user_menu.py
@@ -10,9 +10,9 @@ from telegram.ext import (
     filters,
 )
 
-import db_data
-from db_utils import get_db_connection, hash_password
-import tasks
+from src.core.db import data_access as db_data
+from src.core.db.utils import get_db_connection, hash_password
+from src.core import tasks
 from src.library_bot.states import State
 from src.library_bot.utils import get_user_borrow_limit, normalize_phone_number
 from src.library_bot import keyboards
